feat(Velocity): ability to track velocity of a given object
The Velocity Estimator class takes samples of the position and rotation
of a given Transform over a set number of frames and then uses that
cached data to estimate the object's velocity and angular velocity.

The Velocity Estimator has been added to the UnityXRCameraRig prefab
so it has a basic way of tracking the velocities of the tracked objects
such as the HMD and controllers.