feat(Velocity): report velocity on first active VelocityTracker
The VelocityTrackerProcessor takes an array of VelocityTracker scripts
and exposes the same methods i.e. `GetVelocity()` and
`GetAngularVelocity()` so instead of needing to determine which is
the active VelocityTracker for the active camera rig, the scene
VelocityTracker scripts are added to the VelocityTrackerProcessor
and then the VelocityTrackerProcessor is queried for the relevant
method e.g. `GetVelocity()` and it will return the velocity for the
first active VelocityTracker in the scene.

This will usually mean because other camera rigs are disabled at
runtime except the single active one, then only the active
VelocityTracker script will be found and report the velocities.