feat(Follow): ability to make a transform follow another transform
The Object Follow script enables the ability to set a source
component to be affected by a collection of target components based
on the injected modifier. The modifier included in this commit is
the TransformSourceFollowActiveTarget which will make the source
follow the first active target found.

The following is done in a generic Moment Processor that removes
the need to pepper scripts with their own Unity game loop methods
such as Update, FixedUpdate, etc.

The Moment Processor allows defined moments to be processed at a
specified point in the Unity game loop, ensuring a single loop
is dealt with and additional controls can be applied such as only
processing the moments if the Component is active in the scene.