feat(Follow): multiple target transforms follow single source transform
Another Follow Modifier that works in the opposite way and instead of
getting the source Transform to follow the first active target
Transform, it gets all target Transforms (even if they're disabled) to
follow the source Transform.

This is useful for having a play area alias that all scene CameraRigs
follow around.