feat(Tracking): apply transformations to Transform based on a target
The Transform Modify script enables transformations to take place on
a given source Transform based on the properties of a target Transform.

This means a target Transform can be positioned, rotated and scaled
and then these properties can be applied to another Transform and over
a given duration (to lerp to the new orientation).

This script replicates the concept of teleporting and dash teleporting
as it receives an updated location and then moves the source Transform
(e.g. the play area) to that new position either instantly or over
a given period of time.