feat(Input): unity input event actions
The Unity input event actions extend the relevant data type event based
on the datatype of the input payload. For example, a button press will
only ever be in 2 states of pressed or released so the Unity Button
Action extends a Boolean Action, whereas an axis change can either be
a 1 dimensional axis (such as an analog trigger) or a 2 dimensional
axis (such as a thumbstick) and these implement Float or Vector2
actions respectively.

A prefab has been added to the UnityXRCameraRig that provides default
mappings for the Unity Buttons which should work with any OpenVR
or Oculus controller. No Axis mappings are added as these need to be
added manually into Unity's Input system.