fix(events): ensure all UnityEvents are instantiated to prevent nulls
If UnityEvents aren't instantiated at the point of creation then they
will throw a null reference exception if they are referenced via code
e.g. if trying to add a listener to an event via code.

They don't do this in the editor because the editor instantiates the
event type when a new persistent event is added.