feat(Pointers): pointer concept with straight pointer implementation
The pointer concept is the ability to designate an area of the scene
with a destination. The Straight Pointer implementation of the pointer
casts a straight beam from an origin point (e.g. controller) and stops
when it collides with an object (or reaches the set maximum length).

A pointer knows when it is activated, deactivated, when it starts
hovering (enters) a new object, when it finishes hovering (exits)
an object and knows what coordinates of the object it is hovering
over. These are all cast as events with a special pointer payload
containing information about the pointer target.

The RayCast emitted by Straight Pointer can be customised by utilising
a PhysicsCast object which is a MonoBehaviour component that allows
edit time configuration of a Physics Cast and can then be injected
into the Pointer script. This can allow optional configuration
such as choosing which layers to ignore and whether to ignore
trigger colliders.

A pointer target's validity can also be controlled by a
PolicyExclusionList which allows custom rules to be set up to
determine if a GameObject is a valid or invalid target.

The TransformModify script has been updated to accept the new
TransformData object as it's parameter as it's more versatile than
simply passing a Transform.