feat(Tracking): locate nearest surface in a given direction
The SurfaceLocator class casts a ray from a defined origin in a given
direction and reports the first surface data it finds in a new
SurfaceData type which contains information about the current surface
collision and the previous surface collision.

The SurfaceLocator can be used in conjunction with the TransformModify
to ensure the play area is always snapped to the nearest floor.

A new Vector3 extension has been added to allow the comparison of 2
Vector3 values with a provided threshold of similarity.

A custom PhysicsCast class has been added to allow overriding cast
information within the SurfaceLocator class.

The ability to denote certain surfaces as valid or invalid is achieved
by the new ExclusionRule class which allows to set up certain rules of
when to determine if something should be excluded or not.