feat(Action): determine if surface position has had specified change
The SurfaceChangeAction digests SurfaceData and compares the current
surface position and the previous surface position and checks to see if
the distance between the two positions is within the specified
threshold.

This can be useful to capture SurfaceLocator events and only carry out
actions if the surface difference is within the threshold. A good use
case is only blink the camera if the surface difference is over a
certain step size.