feat(Pointer): solid straight pointer
The pointer concept is the ability to designate an area of the scene
with a destination. The Solid Straight Pointer implementation of the
pointer casts a straight beam from an origin point (e.g. controller)
and stops when it collides with an object (or reaches the set
maximum length).

A pointer knows when it is activated, deactivated, when it starts
hovering (enters) a new object, when it finishes hovering (exits)
an object and knows what coordinates of the object it is hovering
over. These are all cast as events with a special pointer payload
containing information about the pointer target.