feat(Pointer): pointer implementation
The pointer concept is the ability to designate an area of the scene
with a destination. A pointer consists of multiple independent
functionality which is implemented through a component that casts
and returns points that define the pointer beam, one that
generically renders points via game objects and a third component
handles the selection functionality of a pointer.

A pointer knows when it is activated, deactivated, when it starts
hovering (enters) a new object, when it finishes hovering (exits) an
object and knows what coordinates of the object it is hovering over.
These are all cast as events with a special pointer payload
containing information about the pointer target.

The Straight Line Cast implementation casts a straight line from an
origin point (e.g. controller) and stops when it collides with an
object (or reaches the set maximum length). The same applies to the
added Parabolic Line Cast, but it casts a parabolic line instead.

Additionally the introduction of more components that raycast means
the existing `PhysicsCast` class is moved to the new namespace
`Cast`.