fix(Follow): ensure missing references don't throw null exceptions
A number of places were not defended against having null objects
which would cause null exceptions to be thrown if no object was
provided.

Rather than provide a warning, it just fails silently which is
correct as if a target isn't provided then it's not expected to
follow anything. This may be relevant if the target is to be added
later at runtime via code.