feat(Tracking): collision tracking
To track whether a collision between to objects happens Unity offers
so called "Messages" which are defined methods that are not part of
the inheritance chain. This change adds a new component
`CollisionTracker` that allows to forward any of these messages via
UnityEvents. The events differentiate between a trigger and non-
trigger collision and the event data allows retrieving that
information in addition to the collision and the collider.