feat(Extension): add approximately equals check to float type
When doing an equality check on a float, there is an edge case
where the floats do not match due to rounding issues.

The `ApproxEquals` float extension allows a tolerance to be
provided when doing a float equality check.

Any place that uses an `==` for a float has been updated to use
this new method.