fix(Pointer): remove reliance on cast data for invalid state
The ObjectPointer was relying on the `previousPointsCastData` to
determine if the target was no longer valid and was setting this
to null if the cast had exited a valid target.

However, it can never be null otherwise invalid casting doesn't
work. This fix removes the reliance on `previousPointsCastData`
and introduces a simple `isHovering` bool that determines if
the beam is currently over a valid target.