fix(Pointer): ensure exit/deactivate events are called in correct order
The ObjectPointer `Deactivated` event would always be called even if
the pointer component was disabled. This has been fixed by wrapping
the `Deactivate` method with an `isActiveAndEnabled` but also in the
`OnDisable` a specific `Deactivated` event is emitted otherwise the
event would never fire when the component is disabled at runtime.

The payload with the exit event was also incorrect as it was the
current target payload and not the payload of the previous object
that had just been exited. This has been fixed by caching the
points cast data and using the previous value in the exit event.

Also, the hover duration was being reset to 0 before the exit event
was emitted meaning the exit event didn't include the correct hover
duration.