chore(events): prevent events and actions on disabled components
Events should not be emitted if the component emitting the event
is disabled or the GameObject it is on is inactive in the scene.

By default, Unity will only disable Unity loop methods if a component
is disabled or the GameObject is inactive in the scene but any
method that is manually called or called via an event link will
still execute. This may cause unexpected results if a GameObject is
inactive but linked events are still having the same effect as if
the GameObject is active.

The tests have been updated to ensure that inactive GameObjects
do not emit events or process the actions of methods being called.