feat(Follow): split follow modifiers into more distinct chunks
The TransformSourceFollowActiveTarget and TransformTargetsFollowSource
Follow Modifiers have been removed and split down into smaller more
distinct chunks.

Now there is a simple `ModifyTransform` modifier which simply sets the
source transform data to match the target transform data.

The logic of whether to update source to the active target or to update
all targets to the current source has been separated into their own
classes which simply determine that responsibility and take a modifier
as a parameter to choose how to actually modify the transform.

So the logic now is:

 * If source wants to follow the active target by changing transform
   * Apply a ModifyTransform component
   * Apply a SourceFollowActiveTarget component
   * Provide the ModifyTransform component to the appliedModifier
     parameter of the SourceFollowActiveTarget component
 * If the targets want to follow the source by changing transform
   * Apply a ModifyTransform component
   * Apply a TargetFollowSource component
   * Provide the ModifyTransform component to the appliedModifier
     parameter of the TargetFollowSource component

The previous way of determining if the modifier should process the
first active or all targets has now been changed from a bool state
that had an abstract method to receive the information to an enum
that determines how the modifier should be treated. This is more
descriptive and will allow future enhancements if they are required.