feat(Follow): follow based on modifying rigidbody velocity
The `ModifyRigidbodyVelocity` component gets an object to follow
another object by determining the distance between the two objects
and applying the appropriate amount of velocity to the rigidbody
to ensure it follows the object around.

This has the benefit that all movement of the object is handled within
the Unity physics engine at all times.