refactor(Structure): favour List types over arrays for public fields
It is better to use a `List` type than an array for a public field as
a List comes with easier helper methods to add and remove (or clear)
items from the List. This means if someone is trying to update the List
contents via script at runtime, they don't need to mess about with an
array structure to add or remove items.

A new `IEnumerableExtensions` component has been added that adds a new
method to any `IEnumerable` type (e.g. `List`) called `EmptyIfNull`
and will return an instantiated empty `List` if the variable is null.
This will prevent null exceptions when accessing the variable in
situations like `foreach` loops.