chore(events): add arguments and sender to raise methods
Components that offer events should always offer methods to raise
them. The method that raises the event always needs to follow the
naming convention and additionally should always take the arguments
for the event, that is the data and sender (in that order).

This change applies the necessary changes and additionally
simplifies most of the Action implementations to reduce the number
of methods by moving all shared logic into the base class.