chore(events): add arguments and sender to raise methods
Components that offer events should always offer methods to raise
them. The method that raises the event always needs to follow the
naming convention and additionally should always take the arguments
for the event, that is the data and sender (in that order).

This change applies the necessary changes and additionally renames
the event names of `ObjectPointer` since the raise method
`OnBecameVisible` is a "message method" in Unity already. The
pointer prefabs are both updated for this name change.