chore(events): remove sender object from events
The relevant UnityEvents no longer pass a sender object as this causes
all of the methods to require an optional `object` parameter just to
accept these events.

Now there is no sender object, it is not possible in a listener to
determine what object sent the event, but as it is required to have
the script that the event is on to register the against the event in
the first place, then it's implicitly known anyway.

A large refactor of how the Invokes on the events has been undertaken
so they are no longer separated out into `OnAction` methods and the
Invokes are just done inline where required. This cuts down massively
on the `isActiveAndEnabled` calls that were being done and cleans up
the logic to make it more readable.

The Action components have also been simplified so most of the logic
now sits in the base class and the concrete type classes only provide
the type and any simple overrides required.

The logic of the actions has now also changed slightly, now an action
is only Activated when the value is not the default value, and only
Deactivated when the value is the default value.