refactor(Follow): reduce duplicate code in follow modifiers
There were a number of places where the same logic in different
FollowModifier components was being repeated, such as the caching
of the transforms.

This has all now been moved into the base class so the extended
classes can focus on exactly what they're meant to do and not worry
about the caching.