fix(events): ensure digested event data is not mutated
As event data payloads are the same instance that is just mutated each
time the event is emitted, this means any listener that digests the
event should not attempt to mutate the event data otherwise it will
change the original reference.

Therefore, a new EventData payload should be instantiated and be a
new instance to prevent overriding the original data.

This is achieved by a new `Set` method that accepts the `EventData`
as the parameter and a new `Clear` method that resets the data
back to the default values.