chore(Tracking): collision tracker events
The `CollisionTracker` followed Unity's physics message method names
for its event names. This change renames all events to follow the
code conventions and merges the collision and trigger events to only
leave the collision ones. Additionally a new flags enum allows to
select which collision types to emit the events for.