chore(conventions): rename Unity flags attribute
The code conventions insist on naming a C# attribute with
"Attribute" as a suffix. This change renames the `UnityFlag`
attribute to follow the convention and turns it into plural to match
the `System.FlagsAttribute`. Additionally the usage of backticks in
the documentation has been changed to use the proper XML tag `<c>`
instead.