chore(Drawers): use Everything for flags enums
Flags enums are drawn using Unity's GUI for it that allows selecting
multiple values. This GUI allows to easily set all values by
selecting `Everything`. This change ensures all default values for
members that show in the inspector are showing up as `Everything` by
setting all bits to true, which translates to an integer with value
`-1`.