feat(Action): add source actions for action chaining
Chaining actions to send the output of one to be the input of the
next action is a useful functionality and allows more powerful
abstractions. A boolean action emitting on a button press on a
controller can for example be wired up to trigger another action
that in turn represents a grab. This grab action can also be
triggered in any other way making any response to that action a
generic grab handler.

Actions already support this but instead of having to wire up each
event to the next action in the chain this change adds a list of
actions to the base action and subscribes to all of them when the
action is turned on. Once the action is turned off the action
unsubscribes from any source action. This ensures the expected
behavior of doing nothing while the action is turned off.