feat(Tracking): allow chaining collision notifications
To allow interactors like a controller to interact with an unlimited
number of interactables performance has to be taken into
consideration. Instead of all interactables taking part of Unity's
physics notifications and checking whether the thing they collided
with is an interactor they want to work with, this change allows to
forward the collision events from the interactor to the interactable
it's colliding with. To give the receiving side of a forwarded event
all the needed information the sender of the forwarded collision
notification is added to the `EventData`. A check against the
forwarded collision's game object prevents an infinite loop when a
collision would otherwise get forwarded to itself forever.

Since the `CollisionTracker` always took part in the collision
notifications via Unity message methods this change also splits the
collision notification from the collision tracking by introducing a
new superclass that only provides events to get emitted on calling
its methods to do so. The existing tracker now subclasses this new
class and emits the respective event in response to Unity notifying
it about a collision.