feat(Prefabs): provide xr configuration for Unity XR CameraRig
The `[UnityXRCameraRig]` prefab now has a configuration script
attached to it so it can be better configured for XR usage.

It allows for the Unity Tracking Space to be specified and an option
to determine whether to force the fixed timestep to match the refresh
rate of the headset for better physics activity.