feat(Tracking): add custom equality check for collision notifications
The CollisionNotifier has a custom equality check as it only is really
concerned with equality of collisions based on the containing transform
which could be either the transform the collider is on, or a parent
transform which contains the rigidbody for the compound colliders.

A new Collider Extension has been added to allow an easy way of getting
the actual collision source based on whether the collision is a result
of a compound collider and therefore the source is the Transform with
the Rigidbody, or whether the Transform is the same GameObject
component as that which holds the collider.