feat(Collision): processing of collision tracking data into a list
A CollisionList stores a list of collisions that are received from a
Collision Tracker/Notifier and this list can then be processed to
modify the list contents through some included actions such as
`DistanceSortListProcess` which sorts the collisions by which collider
is closest to the source collider.

The `SliceListProcess` takes either the first or last element from the
list and can also emit this data to one of the new List Actions.

A List Action can be executed when a List Process processes the list
and each collision element can be executed against a List Action that
resides on the component being collided with.

The `SetActiveStateListAction` allows a collision on a source object
to tell the target object to toggle the action state of specified
GameObjects.