feat(Collision): track and process active collisions
The `ActiveCollisionsContainer` holds all existing collisions
taking place from a `CollisionNotifier`. It determines when a first
collision has taken place, when any subsequent collisions happen
and when all collisions end.

This `ActiveCollisionsCollection` can then be passed through to other
components that perform operations on the collection and output a new
copy of the collection with any processing applied.

 * `OrderReverser` accepts a collection and reverses the order of the
 collisions in the collection.
 * `NearestSorter` accepts a collection and based on the set source
 transform will sort the collection based on which colliding transforms
 are closest to the set source transform.
 * `Slicer` accepts a collection and slices up the list from the given
 start index for the given length and returns the new sliced collection
 and provides a mechanism for getting the remaining elements that
 were not included in the slice.

All of these processes emit events with the new collection data and can
be chained together by passing the results of one operation into
the method of another operation to create structures like:

`ActiveCollisionsCollection -> NearestSorter -> Slicer(0,1)`

Will result in a sliced collection containing the closest colliding
object to the source (e.g. controller) and will also provide a
collection of all of the other objects that are colliding but are not
the closest.