feat(Pointers): add prefab helper script for easier setup
A new set of pointer prefab helper scripts have been added to the
pointer prefabs. The main User setup script allows easier set up
of the target Transform to get the pointer to follow without needing
to go into the pointer hierarchy and add the transform to follow to
the ObjectFollow script.