feat(Collision): broadcast active collisions to any interested receiver
The ActiveCollisionsBroadcaster broadcasts a message to any
ActiveCollisionsBroadcastReceiver found within the ActiveCollisions set
on the component. It also has the ability to pass over the specified
GameObject that is the overall GameObject that initiated the collision.

This enables a collision to pass over data to what it has collided with
without needing to tightly couple any of the subsequent components.

It makes it possible to create interaction paradigms like touching,
grabbing and using but without the need for any specific logic relating
to those actions.