chore(Collision): rename active collision parent namespace
The Active Collision components now sit within a namespace called
`Active` instead of `Collection` as `Collection` is far too generic
to describe the purpose of the namespace. Anything within the `Active`
namespace is to do with active collisions so being in the
`Tracking.Collision.Active` namespace makes a lot more sense.