refactor(Action): disallow sources list being accessible at runtime
The sources list is now a protected serialized field meaning it can
be populated in the inspector as previous, but it cannot be directly
updated at runtime as this makes no sense because when the sources
list is updated, the actual sources are not subscribed to.

This has been overcome by adding three methods for adding, removing
and clearing the sources at runtime, which handles the subscription
and unsubscription.