feat(Collision): publish active collisions to linked consumers
The ActiveCollisionPublisher pushes itself onto every stored active
collision that has an ActiveCollisionConsumer that consumes the
publisher and contains additional data such as the source container
for the colliding object.

This enables a collider to pass it's state over to any active collision
and provide logic for enabling certain interaction paradigms like
touching, grabbing and using.