feat(Follow): add optional follow offset
The ObjectFollow component now has an optional offset that is a
Transform which offsets the source orientation against the target.

The offset can be limited to any specific transform property like
position, scale, rotation.

The offset can be used to ensure that a source transform follows
a target but has a slight difference, for example if an interacted
object wants to follow an interacting object but be rotated or
positioned slightly differently upon interaction.