chore(Collision): simplify publishing
The ActiveCollisionConsumer component already checks against being
active and enabled in its Consume method. This change therefore
simplifies the calling side to not duplicate that check and instead
leaves it up to the method implementation of Consume instead.

The usage of `notifier` was renamed to the more appropriate
`consumer` with this change, too.