feat(Rule): make rules more flexible and powerful
The existing `ExclusionRule` class was used to allow filtering on
any component that needs it. This change doesn't change the intended
use case, but allows rules to be more powerful.

A new interface `IRule` defines what a rule does: It determines
whether a given object is accepted by a rule. The change from a
negated question "is this excluded" to "is this accepted" makes
reasoning about the rules and their usage easier. To use a rule in
the inspector a `RuleContainer` type is added in addition to a
helper extension method on that type, which allows calling the
`Accepts` method without checking for a `null` rule first. The
existing `ExclusionRule` is split into multiple standalone rules and
some additional rules allow to combine rules in any way possible.
Existing usages in components and tests have been updated and all
new components and types come with unit tests ensuring their
functionality is as intended.

To allow picking a type for the `AnyComponentTypeRule` a new
attribute and property drawer have been added. `RuleContainer`s can
be set and managed right in the inspector via a plus/minus button
that takes care of adding/removing the necessary rule components to
the game object while still allowing the default drag and drop
behavior.