feat(Collection): add key value relation collection for game objects
The GameObjectRelations component allows a List of Relations, which are
a key/value pair of GameObjects to be set up and then when a given
GameObject key is provided it is looked up in the list and the matching
value is returned and emitted as an event.

This can be used for setting up snap handles relevant to the controller
interacting with the current object.