feat(Collection): emit events when game object is push/popped on stack
The GameObjectEventStack component allows a list of events to be
defined that will occur in order when a GameObject is either pushed
(added) or popped (removed) from the stack.

When a GameObject is pushed to the stack, it will invoke the events
associated with the index that matches the current stack index and
will use the pushed GameObject in its payload.

When a GameObject is popped from the stack, it will invoke the
relevant events using the GameObject in its payload. The stack
can be popped at any point and when this occurs, all other items
higher up in the stack are forcefully popped off the stack as well
which emit a different event.

This is useful if when a certain GameObject is received then certain
events are required to be done with it, and then when it is removed
other events are required.

It can be used to simulate multi grab holding of interactions where
the first GameObject received can be used to call off to grabbing
actions, and the second GameObject received can be used to call off
to secondary actions such as direction control.