fix(Follow): rename source to target on follow components
The follow components had an issue where `target` referred to the
object that was used as the source of the data and `source` referred
to the object that was to be modified.

This is backwards as the `target` is the thing that is to be modified
and the `source` is the source data to be used in the modification.