feat(Follow): rework the object follow paradigm
The ObjectFollow component has now been split up so it doesn't manage
the entire position, rotation, scale modification, instead it hands it
off to the FollowModifier component which in turn now has 3 separate
components that handle each of the desired property modifications.

This means the ModifyTransform and ModifyRigidbodyVelocity components
have been removed and split up into the individual Property modifiers.

This makes it cleaner, more separated and reduces the level of
complexity in the follow scripts.

Instead of having different scripts determine when to get the source
to follow the target or vice versa, this is now just an option on the
ObjectFollow component making it easier to just configure from the
inspector.