chore(Structure): ensure any reference to scene objects uses GameObject
There was an inconsistency in a number of scripts where if a reference
to a scene object was required as a parameter then it would sometimes
use `Transform` and other times use `GameObject`. This commit ensures
everything consistently uses `GameObject` and never `Transform` even
if it may seem that `Transform` makes more sense in some places, it's
really a confusing component type because it is simply the properties
of the transform on a GameObject and really the GameObject is the
actual object within the scene that is to be referenced.

The `Transform` can always be referenced from the `GameObject` as
Unity provides that as standard.

This makes it more consistent with events that emit GameObjects as
now any script that requires a GameObject parameter also has a
complimentary method to set and clear the parameter so they can
be set via events.