feat(Event): proxy emitter for events with GameObject payloads
The GameObjectEventProxyEmitter can be used to clean up setup where
a GameObject event may be doing the same thing multiple times such
as in interactions with the GameObjectEventStack.

In that situation, it's possible to have the GameObjectEventStack
events simply pass on the GameObject event to a
GameObjectEventProxyEmitter and then set up duplicate logic in
the single proxy emitter.