feat(Interactions): add prefabs for interactors and interactables
Two new prefab types have been added the Interactor which is added
to an object that wants to be able to interact with an Interactable.

For example, a controller would have the Interactor added to it as
a child of the controller object and then any object that can be
interacted with would be added as a child to a new Interactable
prefab.

This is all that is needed to enable the ability to touch and grab
objects in a scene.