chore(Tracking): make camera rig alias a core component
Rather than have the CameraRig Alias for the UnityXR prefab part of the
prefab structure, it has now been moved to the core Tracking.CameraRig
namespace as it is a generic script that just allows linking up common
CameraRig aliases with the associated collection.

This generic linked alias component can be used for more than just the
UnityXR prefab so it is better for it to reside in the core namespace.