chore(Modification): rename direction controller to direction modifier
The use of the word Controller is misleading as it implies a deeper
level of control and in reality the component is simply modifying the
direction of a source. The DirectionController is now named
DirectionModifier to keep it in line with the naming convention of the
TransformModifier as both scripts simply modify something and there is
little controlling happening.