chore(Modification): rename transform modifier
The TransformModifier has been renamed to TransformPropertyApplier as
it is a better representation of what the component is actually doing.

The component's purpose is to take the properties from a source
transform and apply them to the given target transform, therefore
modifying the target transform properties but through the application
of source properties.

Therefore, it makes more sense to describe the function of the
component as an applier rather than a modifier, because the
modification is a product of the application and the applying of the
properties is the main function of the component.