chore(Modification): move component to appropriate namespace
The GameObjectStateMirror has been moved to Modification as it
actually modifies the state of a GameObject to mirror the state of
another GameObject.

Whilst it is still tracking the state of the source GameObject so
being in Tracking namespace is fine, but it's actually modifying
something as well so Tracking.Modification is more appropriate.