feat(Transformation): transform an input datatype to an output datatype
A Data.Type.Transformation collection has been added that provides
a number of operations that can take a given data type and transform
the input in some way and provide a new output.

For example, being able to take a boolean and output the float
equivalent or to receive a float and multiply it by a known multiplier.