refactor(Interactions): reorganise interactable prefabs
The interactable prefabs have been reorganised so the logic that deals
with grabbing and ungrabbing has been split into more logical chunks
by placing them in their own child GameObjects. This makes it easier
to customise the prefabs by simply turning on or off the GameObjects
to create a customised state.