feat(Transformation): transform float to boolean
The FloatToBoolean DataType Transformer takes a float and if it is
within the specified range will convert it to a true boolean, if the
value is outside of the range it will be converted to false.

A number of additional types and attributes have been added to aid
in defining a range and displaying a range with a simple slider.