feat(Action): allow action sources to be disabled in ActionRegistrar
The ActionRegistar ActionSource now has an enabled flag on the
component and if it is unchecked it will prevent that ActionSource
from being used even if a valid Register is called using the
matching container.

This is useful if a default ActionSource list is set up but needs
customising at runtime without actually deleting elements from the
list.