feat(Event): add rule to GameObjectEventProxyEmitter
The GameObjectEventProxyEmitter now has a validity rule that will be
checked to see if the GameObject received is valid before re-emitting
it.

The structure of the EventProxyEmitters has been refactored so they
all extend a common base class.

A new ActiveCollisionConsumerEventProxyEmitter has been added as well
to proxy the EventData from the ActiveCollisionConsumer.

Tests have also been added for all of these EventProxyEmitters.