chore(Editor): ensure prefab changes are shown
By default Unity shows values that are part of a prefab instance
differently in case that value is changed compared to the prefab's
default value. This change ensures that Unity does the same to all
of the custom drawers.

This change additionally cleans up and fixes the `MinMaxRangeDrawer`
not serializing changes.