feat(Interactions): recreate interactable prefabs to be more versatile
The Interactables have been re-created with a new more generic setup
that no longer relies on the ActionRegistrar for knowing when
controller buttons are being pressed or released. This means that
toggling the grab of an object can be done more generically without
needing to worry about the actions on the controller.

The new GameObjectSet class provides a mechanism for holding a set of
unique GameObjects and is used to determine what is currently touching
or grabbing a given Interactable.

All Interactables now use the GameObjectEvent stack for dealing with
multiple grabs, even a swap hand grab utilises the stack to deal with
the secondary grab event and swapping the hand of the grab accordingly.

The facade for the Interactable has been updated with more appropriate
events and also deals with passing the relevant interaction events onto
the Interactor facade as previously there was no way of knowing
reliably what event data should be passed from the Interactor when it
was interacting with something.

A new InternalSetting attribute has been added as well that can be
applied to a public field and change the appearance in the Unity Editor
to make the fields look as if they shouldn't be edited. This will help
with settings on the Facades that shouldn't be changed by the end user.